move "Warn about duplicates on import" preferences option (#8937)
